mypy: fix all 10 remaining errors, closing out training/neural (and the whole mlframe mypy sweep at 0)
Final batch: 2 optimizer state-setter assignments (SAM/Lookahead both wrap a base
torch.optim.Optimizer and expose a custom state setter accepting a plain dict; torch's
own Optimizer.state stub is narrower -- defaultdict[Tensor, Any] -- than what the
setter legitimately accepts) needed scoped type: ignore[assignment]. The remaining 8
are the same cast(torch.Tensor, ...) / int(...) / np.asarray(...) no-any-return
pattern used throughout this sweep, one per file (Muon/SAM/Lookahead optimizer math,
ranker loss, numerical/categorical embedding heads, history recorder softmax).

Verified: Success: no issues found in 1133 source files now reports zero errors across all 1133 source
files (946 -> 0 over this session's full sweep). All 8 files import cleanly; targeted
optimizer/embedding/history-recorder test suite -- 109 passed, 6 skipped (CUDA-only),
0 failed.

+++ src/mlframe/training/neural/_categorical_embeddings.py
@@ -24,7 +24,7 @@ def default_embed_dim(cardinality: int) -> int:
     Grows sub-linearly with cardinality and caps at 50 so a very-high-card column doesn't blow up the parameter count. ``card`` here is the
     number of distinct categories (NOT counting the reserved unknown row); the floor of 1 keeps a binary / constant column at a usable width.
     """
-    return max(1, min(50, round(1.6 * (max(1, cardinality) ** 0.56))))
+    return int(max(1, min(50, round(1.6 * (max(1, cardinality) ** 0.56)))))
 
 
 class CategoricalEmbedding(nn.Module):
